pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cogito"
version = "0.1.0"
dependencies = [
    "google-generativeai",
    "requests",
    "python-dotenv",
    "aiohttp",
    "openai>=1.0.0",
    "pyyaml",
    "numpy>=2.2.4",
    "anthropic>=0.49.0",
    "groq",
]

[project.scripts]
cogito-thesizer = "src.syncretic_catalyst.thesis_builder:main"